Raise exception when encountering duplicated optimizer state_dict keys (#148)
Given that we might form multiple groups of the parameters in one model
part, would like to add checks to ensure that the optimizers from
different groups have unique keys. This is because we flattened all
optimizer state_dicts in checkpoint saving/loading. So add the check to
ensure it is safe to keep the parameter groups.